Se ha modificado la estructura de archivos se ha refactorizado el codigo
 Changes to be committed:
	new file:   src/backend/microservices/data_cleaning/.env
	new file:   src/backend/microservices/data_cleaning/app/__init__.py
	renamed:    src/backend/microservices/data_cleaning/config.py -> src/backend/microservices/data_cleaning/app/config.py
	new file:   src/backend/microservices/data_cleaning/app/database.py
	renamed:    src/backend/microservices/data_cleaning/main.py -> src/backend/microservices/data_cleaning/app/main.py
	new file:   src/backend/microservices/data_cleaning/app/utils/__init__.py
	renamed:    src/backend/microservices/data_cleaning/data_cleaning_functions.py -> src/backend/microservices/data_cleaning/app/utils/data_cleaning_functions.py
	renamed:    src/backend/microservices/data_cleaning/aux_functions.py -> src/backend/microservices/data_cleaning/app/utils/dataset_utils.py
	renamed:    src/backend/microservices/data_cleaning/load_dataset.py -> src/backend/microservices/data_cleaning/app/utils/load_dataset.py
	new file:   src/backend/microservices/data_cleaning/test/__init__.py
	new file:   src/backend/microservices/data_cleaning/test/test_cleaning.py
	deleted:    src/backend/microservices/dataset_loading/destinations.env
	deleted:    src/backend/microservices/dataset_loading/load_dataset.py
	modified:   src/backend/microservices/dataset_loading/main.py
	modified:   src/backend/microservices/dataset_loading/requirements.txt
