+++ src/backend/microservices/data_cleaning/app/utils/data_cleaning_functions.py
@@ -1,5 +1,4 @@
 # Importar librerias
-
 import pandas as pd
 import numpy as np
 from sklearn.preprocessing import LabelEncoder, MinMaxScaler, StandardScaler
