--- src/backend/microservices/dataset_loading/destinations.env
@@ -1,4 +1,0 @@
-DATA_CLEANING_URL=http://localhost:8001/receive-dataset/
-EDA_ANALYSIS_URL=http://localhost:8002/receive-dataset/
-DATA_MODELING_URL=http://localhost:8003/receive-dataset/
-MODEL_PREDICTION_URL=http://localhost:8004/receive-dataset/
