+++ src/backend/microservices/dataset_loading/main.py
@@ -1,12 +1,6 @@
-from fastapi import FastAPI, File, UploadFile, Form, HTTPException, Depends, Security
-from fastapi.security.api_key import APIKeyHeader
+from fastapi import FastAPI, File, UploadFile, HTTPException
 import shutil
 import os
-import redis
-from io import BytesIO
-import pandas as pd
-import load_dataset as ld
-import pyarrow as pa
 import logging
 from typing import List
 from pydantic import BaseModel
